# [PERF] formulas: add a cache on linear search functions