Benchmark for 10 000 Vlookup performed on a random array composed of integer values ​​between 1 and 10000
`=VLOOKUP(A1:A10000,A1:A10000,1,false)`

measures on "linearSearch" before the commit:
- 267 ms (48% of evaluateAllCells)
- 271 ms (48% of evaluateAllCells)
- 296 ms (50% of evaluateAllCells)

measures on "linearSearch" after the commit:
- 4 ms (2% of evaluateAllCells)
- 1 ms (0.5% of evalauteAllCells)
- 2 ms (3 % of evaluateAllCells)

Task: [4080146](https://www.odoo.com/odoo/2328/tasks/4080146)